analyzer/requirements.txt
google-genai>=0.2.0
watchdog>=3.0.0
python-dotenv>=1.0.0
